Implement XEP-0392: Consistent Color Generation
This adds methods to generate a color for a JID (or nickname or whatever).
It also supports the color corrections for color vision deficiencies.
Only background color correction is not covered by this, but may be
added later.

It uses the hsluv-c library to convert the hue/saturation/lightness
values to RGB colors. hsluv-c is licensed under MIT and is included now.