Fix potential SEGFAULT on connection error
`socketError()` calls `connectToNextDNSHost()` which might cause
`socketError()` synchronously (and recursively), thus not giving a
change for updating `nextSrvRecordIdx`.

Overall, this results in attempting to connect to the same DNS record
recursively, until the stack is exhausted, resulting in SEGFAULT.

One of the solutions (done in this commit) is to increment the record
index _before_ attempting to connect.