Revert "Implement MIX-PAM XEP-0405: Roster IQ extension"
This reverts commit 740a085ef7ac707e2cc2217edf02e296c3f7692e.

There were talks on the standards mailing list that the XEP will be
changed and a private PEP node is used for distributing joined channels.
Also no server (that I am aware of) supports the MIX roster extension,
so I think it is the best we remove before the next release, so we do
not have problems with deprecations and ABI compatibility.