Introduce QXmppInternalClientExtensions
The new internal client extensions can be used to access private parts
of the QXmppClient since QXmppInternalClientExtensions are a friend of
the QXmppClient. The internal client extensions are not part of the
public API, so you can't access them in an application using QXmpp.
However, this was also the case before.

If there is the need to access the internal extensions in the future,
we'll still have all options open to change everything, because the
classes are not part of the public API.