Make QXmppTlsManager an internal client extension
This removes the QXmppTlsManager from the public API and makes it an
internal client extension. It was not of any use for the end user and
was configured via the QXmppClient (as before). This way we can obsolete
some public methods of the QXmppClient, that also have been added with
the new TLS manager.

This can be done without any concerns because the manager was not part
of the public API in any release, yet.