QXmppClient: Remove obsolete methods for TLS management
The methods are obsolete since the TLS manager has become an internal
client extension. The methods were not of any use except for the TLS
manager and should better not be public to the user at all (i.e.
startSocketEncryption()).

This can be done without any concerns because the methods were not part
of the public API in any release, yet.