RosterManager: Cache roster when using stream management
The roster manager caches the roster items if a previous stream could
be resumed now. The rosterReceived() signal is not emitted again in
this case. This avoids redundant roster requests.

When no stream management is used, the roster cache is cleared directly
on disconnecting. If stream management is used, the cache is also
accessible after disconnecting. On connecting the manager checks
whether the last stream could be resumed. If the last stream hasn't
been resumed, the cache is cleared and the roster is re-requested.