Stanza: Make E2eeMetadata optional to avoid allocation
Avoids an allocation of the E2eeMetadataPrivate object when it's not
used.

It doesn't use optional<E2eeMetadata> directly because that wouldn't
avoid the allocation and takes up 16 bytes instead of 8.

This concept could be generalized and used elsewhere in the code
(e.g. extended addresses in the stanza are unused 99% of the time).