Add new message carbons manager
The new manager automatically enabled carbons and the user doesn't need
to do anything.  Messages are injected into the message pipeline of the
client instead of using a signal.  This makes it possible to decrypt the
messages in the next step or do any other possible post-processing on
the message.

Adapting the old manager was not possible because that would mean major
behaviour change that could potentially lead to clients processing
message twice.