Add new QXmpp logo (#528)
This commit adds a new logo for the project. The logotype includes
the alphabet 'Q' written in Morse Code, this is inspired by the use of
morese code in the logotype of a downstream project - Kaidan. The
logotype resembles a chat speech bubble, hinting that this library is
used to facilitate chat communication.