Set errors to be ignored during account creation/deletion
Some servers close the connection during account creation with a
<connection-timeout/> stream error which can be ignored.

https://xmpp.org/extensions/xep-0077.html#usecases-cancel specifies the
usage of the stream error <not-authorized/>.
That is now handled as a valid response to the account deletion request
instead of a result IQ stanza or in addition to it.
The <conflict> stream error is now handled as well since some servers send
it instead of the <not-authorized/> stream error.

The KeepAliveError was emitted during testing between requesting the deletion
of an account and the server's response.
That seems to be due to an XMPP ping that could not be handled appropriately
by the server.
The error is now handled as a response to a successful account deletion.