Set stream errors to be ignored during account creation/deletion
https://xmpp.org/extensions/xep-0077.html#usecases-cancel specifies the
usage of the stream error <not-authorized/>.
That is now handled as a valid response to the account deletion request
instead of a result IQ stanza or in addition to it.
The same applies to the stream error <conflict> since some servers send
that instead of the <not-authorized/> stream error.

There are servers closing the connection during account creation with a
<connection-timeout/> stream error.
That is now ignored.