Stream/Client: IQ handling: Automatically parse stanza errors
This changes the behaviour of the IQ handling. If an IQ response of type
error is received, the stanza error is parsed and returned as
QXmppError. This reduces the amout of further code on side of the user
(and makes the binaries smaller). It makes it impossible to parse the
rest of an error IQ though. However a function for that purpose may
still be added later.