Client: Fix trial of next DNS record on socket errors before connection
When trying different DNS records there are two different error
scenarios:
 1. The error occurrs before TCP connection establishment.
    Example: No service listens on the port.
 2. The error occurrs after TCP connection establishment.
    Example: XMPP server doesn't serve this hostname (stream error)
In the first case we can directly start the connection to the next
record. In the second case we need to wait for the socket to be
disconnected again first.

The code expected that the socket disconnected() signal was emitted in
both cases, but this isn't true for the first case. This commit fixes
the behaviour by checking whether the socket is still connected and a
disconnect signal will be emitted.