TestClient: Allow to disable resetting the id count
Some tests are more complex and involve inner tasks in between calls
and the test client does not handle that great, leading to iq id clash
and random iq id generation which make tests unreliable.
This is especially visible when a test involve many expect/inject calls
like in the migration manager tests.