OutgoingClient: Fix ssl connection/setting invalid ssl config
Constructing a `QSslConfiguration()` result in a null/empty
configuration. The fact that it worked with Qt 5 is pure luck, ie Qt bug.
The bug eventually get fixed in Qt 6 but not backported which lead to
this needed fix. We need to init from the default configuration else we
have no certificates filled.